On RPM old version (<= 4.9), still enable different Python binary install.
On the old version, the RPM source does not have setup.py.in.
In the case such as CentOS6, only way to install RPM Python binding is
by different Python built RPM Python binding RPM.

If RPM Python binding works on CentOS6 now, we should not disable the way
to install for this case.
We should keep the compatibilty.