openSUSE tumbleweed: drop Python 2.7 and change Python 3.7 to Python 3.
Initially we tried to fix the following error on tumbleweed.
We removed python2-tox and added python-xml to run `tox -e py27`.

```
No provider of 'python2-tox' found.
```

But we recognize rpm-python2 package does not exist on tumbleweed.
That causes another error.
https://travis-ci.org/github/junaruga/rpm-py-installer/builds/740486451#L831

While the current RPM version is 4.15.1 on tumbleweed,
it will have the 4.16 dropping Python 2 support soon.
So, we dropped the Python 2.7 case on tumbleweed.

For the Python 3.7 case, we see the following error.
And we decided to simplify to maintain the openSUSE changing py37 to py3.
 until we will see someone who wants to maintain it.

```
ERROR: InterpreterNotFound: python3.7
```

See the pull-request #227 for detail.